Change unsigned long to unsigned long long to handle 64 bit ranges.

+++ kernel/arch/x86/memblock.h
@@ -2,8 +2,8 @@
 #define __MEMBLOCK_H__
 
 struct memblock_region {
-	unsigned long addr;
-	unsigned long size;
+	unsigned long long addr;
+	unsigned long long size;
 };
 
 struct memblock_type {
@@ -17,8 +17,8 @@ struct memblock {
 	struct memblock_type reserved;
 };
 
-int memblock_add(unsigned long addr, unsigned long size);
-int memblock_reserve(unsigned long addr, unsigned long size);
+int memblock_add(unsigned long long addr, unsigned long long size);
+int memblock_reserve(unsigned long long addr, unsigned long long size);
 unsigned long memblock_alloc_range(unsigned long size, unsigned long align,
 				unsigned long from, unsigned long to);
 void memblock_free_range(unsigned long addr, unsigned long size);
